anomaly-agent/pyproject.toml
[project]
name = "anomaly-agent"
version = "0.1.0"
description = "Anomaly Agent for SRE incident detection and Jira integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk==0.2.10",
    "asyncclick>=0.6.2",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langchain-openai >=0.1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "langchain-mcp-adapters",
    "langfuse",
    "langchain-openai",
    "langchain-community>=0.3.0",
    "langgraph==0.5.3",
    "langgraph-prebuilt>=0.0.1",
    "langchain_aws",
    "requests>=2.31.0",
    "langchain-ollama==0.3.6",
    "fastapi>=0.104.0",
    "aiohttp>=3.8.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "structlog>=23.1.0",
    "prometheus-client>=0.17.0",
    "psutil",
    "asyncpg",
    "langchain_litellm",
    "langchain_groq",
    "redis",
    "grandalf"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 
